+++ changelog.md
@@ -2,6 +2,9 @@
   - No global stats collection by default.
     There is a ets table based stats collector to record the number of sent bytes and messages. Consider this feature deprecated.
     Since 1.7.0, there there is a better integration for metrics.
+  - Allow to use caller owned ETS table for `wolff_producers` to store partition worker processes.
+    This should avoid having to create an atom for each supervised producer.
+
 * 1.8.0
   - Add wolff:check_if_topic_exists/2 for checking if a topic exists making use of an existing client process. [#52](https://github.com/kafka4beam/wolff/pull/52)
   - Improved logs when reporting connection errors. (merged 1.5.12)